+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.10.2
 
 - Fix: expose x start/end and y start/end for horizontal and vertical line annotations.
+- Fix: ensure KDBush worker works in prod build
 
 ## 1.10.1